DB-18 added order and order status controllers and some changes

+++ src/controllers/orderStatusControllers/deleteOrderStatus.js
@@ -2,6 +2,7 @@ import { config } from "dotenv";
 import responsse from "../../../responsse.js";
 import OrderStatus from "../../models/orderStatusSchema.js";
 import { ResponseHandler } from "../../utils/responseHandler.js";
+import Order from "../../models/orderSchema.js";
 
 const LANG = config(process.cwd, ".env").parsed.LANG;
 
@@ -21,13 +22,18 @@ const deleteOrderStatus = async (req, res) => {
   }
 
   try {
-    const orderStatus = await OrderStatus.findByIdAndDelete(orderStatusId);
+    Order.findOne({ orderStatus: orderStatusId }).then(async (data) => {
+      if (data) {
+        return ResponseHandler(res, "orderStatus", 404, lang);
+      }
+      const orderStatus = await OrderStatus.findByIdAndDelete(orderStatusId);
 
-    if (!orderStatus) {
-      return ResponseHandler(res, "orderStatus", 404, lang);
-    }
+      if (!orderStatus) {
+        return ResponseHandler(res, "orderStatus", 407, lang);
+      }
 
-    return ResponseHandler(res, "common", 203, lang);
+      return ResponseHandler(res, "common", 203, lang);
+    });
   } catch (error) {
     return ResponseHandler(res, "common", 500, lang);
   }
